Makefile: added a deploy target
Right now only I can push the container, but as soon as
other users log in at https://registry.huc.knaw.nl/ for the first time,
I think I should be able to give them rights to push.

The idea is that eventually a push there will trigger a CD pipeline.
That however, is NOT set up yet. The k8s deployments needs to be
restarted explicitly for now in addition to this step.