added an extra 'textannodata' JSON-LD context and tiny ontology
This context defines aliases that map various properties to DMCI,
schema.org and some others. These are to be used in the bodies of
annotations. The aliases are documented in `text_intro.md`.

Terms from TEI, PageXML and other sources will have to
be explicitly mapped to a fitting category. This effort is by no means
complete yet, but this is an indication of the direction I'd like to go
in.

Highly project-specific properties are not included here, these will
land in separate context definitions/ontologies.